Move `does_agent_need_repath` to `mod agent`.
This isn't really necessary, but it makes `lib.rs` shorter, and this code mostly only matters for agents.
